deckbot
Deck of cards bot, intended for use specifically as a bot for a DnD campaign. Deck images not included in the commit because I forgot, might add them later but its not important. Had ideas for other features but the campaign it was for got cancelled so for now its just going to be uploaded.